Edit user form changes, list of private messages rebuilt

+++ app/controllers/private_messages_controller.rb
@@ -2,10 +2,14 @@
 class PrivateMessagesController < ApplicationController
 
   before_filter :authenticate_user!
-
+  
   def index
-    @received_messages = current_user.received_messages.order_by([:created_at, :desc]).limit(50)
-    @sent_messages = current_user.sent_messages.limit(20)
+    per_page = 5
+    if params[:type] == :sent
+      @messages = current_user.received_messages.order_by([:created_at, :desc]).page(params[:page]).per(per_page)
+    else
+      @messages = current_user.sent_messages.order_by([:created_at, :desc]).page(params[:page]).per(per_page)
+    end
   end
 
   def show
@@ -28,10 +32,28 @@ class PrivateMessagesController < ApplicationController
   end
 
   def destroy_multiple
-    messages = PrivateMessage.where(:_id.in => params[:private_message_ids])
-    puts messages.count
-    messages.each {|m| m.destroy if m.receiver == current_user}
+    unless params[:private_message_ids].nil?
+      messages = PrivateMessage.where(:_id.in => params[:private_message_ids])
+      messages.each {|m| m.destroy if m.receiver == current_user}
+    end
     redirect_to private_messages_path
+  end
+  
+  def mark_as_read_multiple
+    unless params[:private_message_ids].nil?
+      messages = PrivateMessage.where(:_id.in => params[:private_message_ids])
+      messages.each {|m| m.read! if m.receiver == current_user}
+    end
+    redirect_to private_messages_path
+  end
+  
+  def destroy
+    @message = PrivateMessage.find(params[:id])
+    if @message.receiver == current_user and @message.destroy
+      redirect_to private_messages_path, :notice => "Wiadomość została usunięta"
+    else
+      redirect_to private_messages_path, :error => "Wiadomość nie została usunięta"
+    end
   end
 
 end
